gateway/requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.0
bedrock-agentcore-starter-toolkit>=0.1.0
